Revert "Se agrega mapper y se crea el service para la generacion del token"
This reverts commit 562e49a617fded9cecde2949f770b450d311f700.
