--- src/main/java/com/macons/apparkme/entities/User.java
@@ -1,14 +1,0 @@
-package com.macons.apparkme.entities;
-
-import lombok.*;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Setter
-@Getter
-@Builder
-public class User {
-
-    private String user;
-    private String password;
-}
