--- src/main/java/com/macons/apparkme/service/TokenService.java
@@ -1,5 +1,0 @@
-package com.macons.apparkme.service;
-
-public interface TokenService {
-    String getJWTToken(String username);
-}
