--- src/main/java/com/macons/apparkme/service/impl/TokenServiceImpl.java
@@ -1,61 +1,0 @@
-package com.macons.apparkme.service.impl;
-
-import com.macons.apparkme.service.TokenService;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import lombok.extern.slf4j.Slf4j;
-import lombok.var;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.stereotype.Service;
-
-import javax.xml.bind.DatatypeConverter;
-import java.security.Key;
-import java.security.SecureRandom;
-import java.util.Base64;
-import java.util.Date;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Slf4j
-@Service
-public class TokenServiceImpl implements TokenService {
-    @Value("${jwt.token.secret}")
-    private String secret;
-    @Override
-    public String getJWTToken(String username) {
-        log.info("{}", generateSafeToken());
-
-        byte[] apiKeySecretBytes = DatatypeConverter.parseBase64Binary(this.secret);
-        Key signingKey = Keys.hmacShaKeyFor(apiKeySecretBytes);
-
-        log.info("\n\n Creando token.... \n\n");
-        String secretKey = "mySecretKey";
-        List<GrantedAuthority> grantedAuthorities = AuthorityUtils
-                .commaSeparatedStringToAuthorityList("ROLE_USER");
-        log.info("\n\n Generando Jwts.... \n\n");
-        String token = Jwts
-                .builder()
-                .setId("apparkmeJWT")
-                .setSubject(username)
-                .claim("authorities",
-                        grantedAuthorities.stream()
-                                .map(GrantedAuthority::getAuthority)
-                                .collect(Collectors.toList()))
-                .setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis() + 600000))
-                .signWith(signingKey).compact();
-
-        return "Bearer " + token;
-    }
-
-    private String generateSafeToken() {
-        SecureRandom random = new SecureRandom();
-        byte[] bytes = new byte[36]; // 36 bytes * 8 = 288 bits, a little bit more than
-        // the 256 required bits
-        random.nextBytes(bytes);
-        var encoder = Base64.getUrlEncoder().withoutPadding();
-        return encoder.encodeToString(bytes);
-    }
-}
